backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.118.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
numpy==2.3.3
pandas==2.3.3
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.37.0
matplotlib==3.8.0
seaborn==0.13.0
langchain==0.3.0
langchain-experimental==0.3.0
langchain-google-genai==2.0.5
python-dotenv==1.0.0
